update:: login, register, find_id,pass, LoginController

+++ hotplace_boot-1/src/main/webapp/script/login.js
@@ -2,25 +2,39 @@ $(function() {
     
     $('#login_button').click(function(){
         if($('#userId').val()=='')  {
-              alert('Please enter user ID');
+	        	Swal.fire({
+		  			  icon: 'error',
+		  			  title: '로그인 오류',
+		  			  text: 'ID를 입력해주세요.'
+				})
               $('#userId').focus();
         }
+        else if($('#userPw').val()==''){
+	        	Swal.fire({
+		  			  icon: 'error',
+		  			  title: '로그인 오류',
+		  			  text: '비밀번호를 입력해주세요.'
+				})
+        }
         else {
-             /* $('form').submit();*/
-        
-        
-       
-        $.ajax({
-        	type:'post',
-        	url:'login.do',
-        	data:"userId="+$('#userId').val()+"&userPw="+$('#userPw').val(),
-      
-        	success:function(result){
-        		console.log(result);
-        		localStorage.setItem("loginUser",JSON.stringify(result));
-        		
-        		location.replace('main.html');
-        		
+	        $.ajax({
+	        	type:'post',
+	        	url:'login.do',
+	        	data:"userId="+$('#userId').val()+"&userPw="+$('#userPw').val(),
+	      
+	        	success:function(result){
+	        		if(result.userId!=null){
+	        			localStorage.setItem("loginUser",JSON.stringify(result));
+	        			location.replace('main.html');
+	        		}else{
+	        			Swal.fire({
+	      	  			  icon: 'error',
+	      	  			  title: '로그인 오류',
+	      	  			  text: 'ID 혹은 패스워드가 올바르지 않습니다.'
+	      			})
+	        		}
+	        			
+	        		
         	}
         
         
@@ -29,10 +43,6 @@ $(function() {
         }
         
     })
-    
-    
-    
-    
     
     $('#notuserstartbtn').click(function(){
     	location.replace('main.html');
